requirements.txt
python-dotenv
playwright
pandas
openpyxl

System used: Ubuntu

Commands:
  - sudo apt update && sudo apt upgrade -y
  - python3 -m venv venv
  - source venv/bin/activate
  - pip install -r requirements.txt
  - python3 -m playwright install